requirements.txt
# YouTube Automation System Dependencies

# Supabase
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Video processing (existing dependencies)
edge-tts>=7.0.0
moviepy>=2.0.0
openai-whisper>=20250625

# AI Script Generation
openai>=1.0.0
anthropic>=0.18.0  # For Claude API

# YouTube API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Utilities
requests>=2.31.0

